requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile requirements/dev.in
#
asgiref==3.5.0
    # via
    #   -r requirements/base.txt
    #   django
attrs==21.4.0
    # via pytest
certifi==2021.10.8
    # via
    #   -r requirements/base.txt
    #   requests
charset-normalizer==2.0.12
    # via
    #   -r requirements/base.txt
    #   requests
coverage==6.3.2
    # via -r requirements/dev.in
dj-database-url==0.5.0
    # via -r requirements/base.txt
django==3.2.12
    # via
    #   -r requirements/base.txt
    #   django-filter
    #   django-phonenumber-field
    #   djangorestframework
django-filter==21.1
    # via -r requirements/base.txt
django-getenv==1.3.2
    # via -r requirements/base.txt
django-phonenumber-field[phonenumbers]==6.1.0
    # via -r requirements/base.txt
djangorestframework==3.13.1
    # via -r requirements/base.txt
drf-writable-nested==0.6.3
    # via -r requirements/base.txt
factory-boy==3.2.1
    # via -r requirements/dev.in
faker==13.3.4
    # via factory-boy
flake8==4.0.1
    # via -r requirements/dev.in
idna==3.3
    # via
    #   -r requirements/base.txt
    #   requests
importlib-metadata==4.2.0
    # via
    #   flake8
    #   pluggy
    #   pytest
iniconfig==1.1.1
    # via pytest
mccabe==0.6.1
    # via flake8
packaging==21.3
    # via pytest
phonenumbers==8.12.46
    # via
    #   -r requirements/base.txt
    #   django-phonenumber-field
pluggy==1.0.0
    # via pytest
psycopg2-binary==2.9.3
    # via -r requirements/base.txt
py==1.11.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pyjwt==2.3.0
    # via
    #   -r requirements/base.txt
    #   twilio
pyparsing==3.0.7
    # via packaging
pytest==7.1.1
    # via pytest-django
pytest-django==4.5.2
    # via -r requirements/dev.in
python-dateutil==2.8.2
    # via faker
pytz==2022.1
    # via
    #   -r requirements/base.txt
    #   django
    #   djangorestframework
    #   twilio
requests==2.27.1
    # via
    #   -r requirements/base.txt
    #   twilio
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via
    #   -r requirements/base.txt
    #   django
tomli==2.0.1
    # via pytest
twilio==7.8.0
    # via -r requirements/base.txt
typing-extensions==4.1.1
    # via
    #   -r requirements/base.txt
    #   asgiref
    #   faker
    #   importlib-metadata
urllib3==1.26.9
    # via
    #   -r requirements/base.txt
    #   requests
zipp==3.8.0
    # via importlib-metadata
